requirements.txt
torch
numpy
tyro
diffusers
accelerate
einops
gradio
imageio
imageio-ffmpeg
lpips
matplotlib
omegaconf
open_clip_torch
packaging
Pillow
pygltflib
rembg[gpu,cli]
rich
safetensors
scikit-image
scikit-learn
scipy
tqdm
transformers
trimesh
kiui >= 0.2.3
xatlas
roma
plyfile